Retro Movement Toggle
There are now two different modes of camera movement: Retro and Modern
Retro uses A and D / Left and Right to move the camera. The player cannot sidestep in this mode.
Modern uses the mouse to move the camera left and right. The player has the ability to sidestep in this mode.